Insert data

See merge request dawson-csy3-24-25/511/section2/Eliana-Maara-Hai-David/finalproject-readingapp!42

+++ app/src/main/java/com/example/bookreadingapp/data/entities/Chapters.kt
@@ -20,7 +20,7 @@ class Chapters {
     @PrimaryKey(autoGenerate=true)
 
     @ColumnInfo(name="chapterId")
-    var id: Int = 0
+    var id: Long = 0
 
     @ColumnInfo(name="title")
     var title: String = ""
@@ -28,10 +28,10 @@ class Chapters {
     @ColumnInfo(name="chapterPosition")
     var position: Int = 0
 
-    var bookId: Int = 0
+    var bookId: Long = 0
 
     constructor()
-    constructor(title: String, position: Int, bookId: Int) {
+    constructor(title: String, position: Int, bookId: Long) {
         this.title = title
         this.position = position
         this.bookId = bookId
