+++ app/src/main/java/com/example/bookreadingapp/data/entities/Image.kt
@@ -19,9 +19,9 @@ import androidx.room.PrimaryKey
 class Image {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "imageId")
-    var id: Int = 0
+    var id: Long = 0
 
-    var chapterId: Int = 0
+    var chapterId: Long = 0
 
     @ColumnInfo(name = "imagePosition")
     var position: Int = 0
