+++ app/src/main/java/com/example/bookreadingapp/data/entities/Paragraphs.kt
@@ -20,9 +20,9 @@ import androidx.room.PrimaryKey
 class Paragraphs {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "paragraphId")
-    var id: Int = 0
+    var id: Long = 0
 
-    var chapterId: Int = 0
+    var chapterId: Long = 0
 
     @ColumnInfo(name="paragraphPosition")
     var position: Int = 0
@@ -30,7 +30,8 @@ class Paragraphs {
     var text: String = ""
 
     constructor()
-    constructor(text: String, chapterId: Int, position: Int) {
+
+    constructor(text: String, chapterId: Long, position: Int) {
         this.text = text
         this.chapterId = chapterId
         this.position = position
